+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti/majorz.py
@@ -36,7 +36,7 @@ def parse_election(line, errors):
         return mandates
 
 
-def parse_election_result(line, errors, municipalities):
+def parse_election_result(line, errors, entities):
     try:
         group = line.einheitbez.strip()
         entity_id = int(line.bfs or 0)
@@ -67,12 +67,14 @@ def parse_election_result(line, errors, municipalities):
             raise ValueError()
 
     except ValueError:
-        errors.append(_("Invalid municipality values"))
+        errors.append(_("Invalid entity values"))
     else:
-        if entity_id not in municipalities:
+        if entity_id not in entities and group.lower() == 'auslandschweizer':
+            entity_id = 0
+
+        if entity_id and entity_id not in entities:
             errors.append(_(
-                "municipality id ${id} is unknown",
-                mapping={'id': entity_id}
+                _("${name} is unknown", mapping={'name': entity_id})
             ))
         else:
             return ElectionResult(
@@ -123,13 +125,13 @@ def parse_candidates(line, errors):
     return results
 
 
-def import_file(municipalities, election, file, mimetype,
+def import_file(entities, election, file, mimetype,
                 elected_file=None, elected_mimetype=None):
     errors = []
     candidates = {}
     results = []
 
-    # This format has one municipality per line and every candidate as row
+    # This format has one entity per line and every candidate as row
     csv, error = load_csv(file, mimetype, expected_headers=HEADERS)
     if error:
         errors.append(error)
@@ -140,7 +142,7 @@ def import_file(municipalities, election, file, mimetype,
 
             # Parse the line
             mandates = parse_election(line, line_errors)
-            result = parse_election_result(line, line_errors, municipalities)
+            result = parse_election_result(line, line_errors, entities)
             if result:
                 for candidate, c_result in parse_candidates(line, line_errors):
                     candidate = candidates.setdefault(